backend/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-environ>=0.11.2

# Async Tasks
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Machine Learning
scikit-learn>=1.3.2
pandas>=2.1.3
numpy>=1.26.2
joblib>=1.3.2

# Utilities
python-decouple>=3.8
python-dateutil>=2.8.2

# Development
pytest>=7.4.3
pytest-django>=4.7.0
factory-boy>=3.3.0
faker>=20.1.0
